Implementation of BounTI with watershed logic
Updated implementation of the BounTI concept. Instead of checking the connected voxels to each segment individually, the segments are expanded together and one surrounding voxel is checked iteratively 5 times for each threshold. The algorithm speed in this version is not affected by the number of segments and watershed logic is applied at the boundaries instead of the the largest components assuming priority in each threshold iteration.